+++ packages/working_router_lint/lib/main.dart
@@ -2,6 +2,8 @@ import 'package:analysis_server_plugin/plugin.dart';
 import 'package:analysis_server_plugin/registry.dart';
 
 import 'src/assists/remove_location_tree_element.dart';
+import 'src/assists/wrap_with_location.dart';
+import 'src/assists/wrap_with_multi_shell.dart';
 import 'src/assists/wrap_with_scope.dart';
 import 'src/assists/wrap_with_shell.dart';
 
@@ -14,6 +16,8 @@ class _WorkingRouterPlugin extends Plugin {
   @override
   void register(PluginRegistry registry) {
     registry.registerAssist(RemoveLocationTreeElement.new);
+    registry.registerAssist(WrapWithLocation.new);
+    registry.registerAssist(WrapWithMultiShell.new);
     registry.registerAssist(WrapWithScope.new);
     registry.registerAssist(WrapWithShell.new);
   }